Moved propagate code from Skeleton.cs to Bone.cs
