+++ Ktisis/Overlay/Skeleton.cs
@@ -151,13 +151,13 @@ namespace Ktisis.Overlay {
 							Interop.Alloc.SetMatrix(transform, matrix);
 
 							// handles parenting
-							PropagateChildren(bone, transform, initialPos, initialRot);
+							bone.PropagateChildren(transform, initialPos, initialRot);
 
 							// handles linking
 							if (boneName.EndsWith("_l") || boneName.EndsWith("_r")) {
 								var siblingBone = bone.GetMirrorSibling();
 								if (siblingBone != null)
-									PropagateSibling(siblingBone, transform->Rotation.ToQuat() / initialRot);
+									siblingBone.PropagateSibling(transform->Rotation.ToQuat() / initialRot);
 							}
 						}
 
@@ -169,49 +169,6 @@ namespace Ktisis.Overlay {
 			}
 		}
 
-		private unsafe static void PropagateChildren(Bone parent, hkQsTransformf* transform, Vector3 initialPos, Quaternion initialRot) {
-			// Bone parenting
-			// Adapted from Anamnesis Studio code shared by Yuki - thank you!
-
-			if (!Ktisis.Configuration.EnableParenting)
-				return;
-
-			var sourcePos = transform->Translation.ToVector3();
-			var deltaRot = transform->Rotation.ToQuat() / initialRot;
-			var deltaPos = sourcePos - initialPos;
-
-			var descendants = parent.GetDescendants();
-			foreach (var child in descendants) {
-				var access = child.AccessModelSpace(PropagateOrNot.DontPropagate);
-
-				var offset = access->Translation.ToVector3() - sourcePos;
-				offset = Vector3.Transform(offset, deltaRot);
-
-				var matrix = Interop.Alloc.GetMatrix(access);
-				matrix *= Matrix4x4.CreateFromQuaternion(deltaRot);
-				matrix.Translation = deltaPos + sourcePos + offset;
-				Interop.Alloc.SetMatrix(access, matrix);
-			}
-		}
-		private unsafe static void PropagateSibling(Bone sibling, Quaternion deltaRot) {
-			if (Ktisis.Configuration.SiblingLink == SiblingLink.None) return;
-
-			var access = sibling.AccessModelSpace(PropagateOrNot.DontPropagate);
-			var offset = access->Translation.ToVector3();
-
-			if(Ktisis.Configuration.SiblingLink == SiblingLink.RotationMirrorX)
-				deltaRot = new(-deltaRot.X, deltaRot.Y, deltaRot.Z, -deltaRot.W);
-
-			var matrix = Interop.Alloc.GetMatrix(access);
-			matrix *= Matrix4x4.CreateFromQuaternion(deltaRot);
-			matrix.Translation = offset;
-
-			var initialRot = access->Rotation.ToQuat();
-			var initialPos = access->Translation.ToVector3();
-			Interop.Alloc.SetMatrix(access, matrix);
-
-			PropagateChildren(sibling, access, initialPos, initialRot);
-		}
 		public unsafe static Bone? GetSelectedBone() {
 			if (!BoneSelect.Active) return null;
